refactor(phase-aj): split daemon dispatch routes
